src/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-storage-file-share
azure-ai-formrecognizer
azure-storage-blob
azure-functions
aiohttp
uuid